@@ -1,6 +1,6 @@
 {
     "name": "eks-example",
-    "version": "0.1.0",
+    "version": "0.1.3",
     "main": "lib/index.js",
     "types": "lib/index.d.ts",
     "bin": {
@@ -12,15 +12,15 @@
         "cdk": "cdk"
     },
     "devDependencies": {
-        "@types/node": "^8.9.4",
-        "typescript": "^3.1.2",
-        "aws-cdk": "^0.19.0"
+        "@types/node": "^13.7.1",
+        "typescript": "^3.7.5",
+        "aws-cdk": "^1.23.0"
     },
     "dependencies": {
-        "@aws-cdk/aws-ec2": "^0.19.0",
-        "@aws-cdk/aws-iam": "^0.19.0",
-        "@aws-cdk/aws-autoscaling": "^0.19.0",
-        "@aws-cdk/aws-eks": "^0.19.0",
-        "@aws-cdk/cdk": "^0.19.0"
+        "@aws-cdk/aws-ec2": "^1.23.0",
+        "@aws-cdk/aws-iam": "^1.23.0 ",
+        "@aws-cdk/aws-autoscaling": "^1.23.0",
+        "@aws-cdk/aws-eks": "^1.23.0",
+        "@aws-cdk/core": "1.23.0"
     }
 }
